pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "rally-segmenter"
version = "0.1.0"
description = "Tennis rally segmentation engine"
requires-python = ">=3.9"
dependencies = [
  "numpy",
  "opencv-python",
  "librosa",
  "soundfile",
  "pyyaml",
  "tqdm",
]

[project.scripts]
rally-segment = "trc.cli.segment:main"
rally-eval    = "trc.cli.eval:main"

[tool.setuptools]
packages = ["trc"]
